refactor: split homepage into separate components
